server/requirements.txt
email-validator
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pandas
numpy
scikit-learn
xgboost
shap
redis
httpx
pytest
